Add vadmin command for non simulation utilities
