+++ src/vivarium_cluster_tools/vadmin/utilities.py
@@ -1,0 +1,18 @@
+import sys
+
+from loguru import logger
+
+
+def add_logging_sink(sink, verbose, colorize=False, serialize=False):
+    message_format = ('<green>{time:YYYY-MM-DD HH:mm:ss.SSS}</green> | '
+                      '<cyan>{name}</cyan>:<cyan>{function}</cyan>:<cyan>{line}</cyan> '
+                      '- <level>{message}</level>')
+    if verbose == 0:
+        logger.add(sink, colorize=colorize, level="INFO", format=message_format, serialize=serialize)
+    elif verbose >= 1:
+        logger.add(sink, colorize=colorize, level="DEBUG", format=message_format, serialize=serialize)
+
+
+def configure_master_process_logging_to_terminal(verbose):
+    logger.remove()  # Clear default configuration
+    add_logging_sink(sys.stdout, verbose, colorize=True)
